+++ lilly/__init__.py
@@ -1,6 +1,6 @@
 """Entry point for the lilly package"""
 
-# Version of the realpython-reader package
-__version__ = "0.1.0"
+# Version of the lilly package
+__version__ = "0.2.0"
 
 from .applications import Lilly