+++ lilly/repositories/sqlalchemy.py
@@ -6,7 +6,7 @@ from typing import Any, Type, Dict, List, Optional
 
 from pydantic import BaseModel
 from sqlalchemy.orm import DeclarativeMeta, Session
-from sqlalchemy import insert, select, update, delete
+from sqlalchemy import insert, select, update, delete, text
 
 from .base import Repository
 from lilly.datasources import SQLAlchemyDataSource
@@ -38,7 +38,7 @@ class SQLAlchemyRepository(Repository):
     def _get_one(self, datasource_connection: Session, record_id: Any, **kwargs) -> DeclarativeMeta:
         return datasource_connection.query(self._model_cls).get(record_id)
 
-    def create_many(self, records: List[Any], **kwargs) -> List[Any]:
+    def create_many(self, records: List[BaseModel], **kwargs) -> List[Any]:
         """
         method to create many records.
         This will always return an empty list as it is quite impossible to return the multiple records inserted
@@ -47,7 +47,7 @@ class SQLAlchemyRepository(Repository):
         This should be overridden.
 
         Arguments:
-            records (list[Any]): the records to be created
+            records (list[BaseModel]): the records to be created
             **kwargs (Any): any extra key-word arguments you may need to pass in your particular implementation
 
         Returns:
@@ -61,7 +61,9 @@ class SQLAlchemyRepository(Repository):
                   skip: int = 0,
                   limit: Optional[int] = None,
                   **filters) -> List[DeclarativeMeta]:
-        query = datasource_connection.query(self._model_cls).filter(*criterion).filter_by(**filters).offset(skip)
+        coerced_criteria = self.__coerce_string_criteria(*criterion)
+
+        query = datasource_connection.query(self._model_cls).filter(*coerced_criteria).filter_by(**filters).offset(skip)
 
         if limit is not None:
             query = query.limit(limit)
@@ -69,42 +71,46 @@ class SQLAlchemyRepository(Repository):
         data = query.all()
         return data
 
-    def _create_one(self, datasource_connection: Session, record: Dict[str, Any], **kwargs) -> DeclarativeMeta:
-        record = self._model_cls(**record)
+    def _create_one(self, datasource_connection: Session, record: BaseModel, **kwargs) -> DeclarativeMeta:
+        record = self._model_cls(**record.dict())
         datasource_connection.add(record)
         datasource_connection.commit()
         return record
 
-    def _create_many(self, datasource_connection: Session, records: List[Dict[str, Any]], **kwargs) -> List[
+    def _create_many(self, datasource_connection: Session, records: List[BaseModel], **kwargs) -> List[
         DeclarativeMeta]:
         stmt = insert(self._model_cls)
 
-        datasource_connection.execute(stmt, records)
+        datasource_connection.execute(stmt, [record.dict() for record in records])
         datasource_connection.commit()
 
         # return empty because it is currently quite hard to return the inserted records with their data
         # unless one is to use the inefficient method of multiple calls to session.add()
         return []
 
-    def _update_one(self, datasource_connection: Any, record_id: Any, new_record: Dict[str, Any],
+    def _update_one(self, datasource_connection: Any, record_id: Any, new_record: BaseModel,
                     **kwargs) -> DeclarativeMeta:
         record = datasource_connection.query(self._model_cls).get(record_id)
         if record is None:
             raise KeyError(f"record with id: {record_id} not found")
 
-        for field, value in new_record.items():
+        for field, value in iter(new_record):
             setattr(record, field, value)
 
         datasource_connection.add(record)
         datasource_connection.commit()
         return record
 
-    def _update_many(self, datasource_connection: Session, new_record: Dict[str, Any], *criterion, **filters) -> List[
+    def _update_many(self, datasource_connection: Session, new_record: BaseModel, *criterion, **filters) -> List[
         DeclarativeMeta]:
-        stmt = update(self._model_cls).filter(*criterion).filter_by(**filters).values(new_record)
+        coerced_criteria = self.__coerce_string_criteria(*criterion)
+        new_data = new_record.dict()
+
+        stmt = update(self._model_cls).filter(*coerced_criteria).filter_by(**filters).values(new_data) \
+            .execution_options(synchronize_session="fetch")
 
         affected_records = self._get_affected_records(
-            datasource_connection, *criterion, new_data=new_record, **filters)
+            datasource_connection, *coerced_criteria, new_data=new_data, **filters)
 
         datasource_connection.execute(stmt)
         datasource_connection.commit()
@@ -120,9 +126,12 @@ class SQLAlchemyRepository(Repository):
         return record
 
     def _remove_many(self, datasource_connection: Session, *criterion, **filters) -> List[DeclarativeMeta]:
-        stmt = delete(self._model_cls).filter(*criterion).filter_by(**filters)
+        coerced_criteria = self.__coerce_string_criteria(*criterion)
+
+        stmt = delete(self._model_cls).filter(*coerced_criteria).filter_by(**filters) \
+            .execution_options(synchronize_session="fetch")
         affected_records = self._get_affected_records(
-            datasource_connection, *criterion, **filters)
+            datasource_connection, *coerced_criteria, **filters)
 
         datasource_connection.execute(stmt)
         datasource_connection.commit()
@@ -146,6 +155,16 @@ class SQLAlchemyRepository(Repository):
             return [_update_orm_instance(record, new_data=new_data) for record in affected_records]
 
         return affected_records
+
+    @staticmethod
+    def __coerce_string_criteria(*criteria):
+        """
+        Wraps any string criteria for filtering in a text()
+
+        e.g.::
+         "title LIKE '%t'" becomes text("title LIKE '%t'")
+        """
+        return [text(criterion) if isinstance(criterion, str) else criterion for criterion in criteria]
 
 
 def _update_orm_instance(orm_instance: DeclarativeMeta, new_data: Dict[str, Any]):