+++ lilly/stubs/service/actions.py
@@ -7,9 +7,23 @@ import string
 
 from pydantic import BaseModel
 
-from lilly.actions import Action
+from lilly.actions import (
+    Action,
+    ReadOneAction,
+    ReadManyAction,
+    CreateOneAction,
+    CreateManyAction,
+    UpdateOneAction,
+    UpdateManyAction,
+    DeleteOneAction,
+    DeleteManyAction,
+)
+from lilly.repositories import Repository
 
+from .dtos import NameCreationRequestDTO
 from .repositories import NamesRepository
+
+_names_repo = NamesRepository()
 
 
 class GenerateRandomName(Action):
@@ -18,7 +32,6 @@ class GenerateRandomName(Action):
     """
     _vowels = "aeiou"
     _consonants = "".join(set(string.ascii_lowercase) - set("aeiou"))
-    _name_repository = NamesRepository()
 
     def __init__(self, length: int = 7):
         self._length = length
@@ -26,7 +39,7 @@ class GenerateRandomName(Action):
     def run(self) -> BaseModel:
         """Actual method that is run"""
         name = self._generate_random_word()
-        return self._name_repository.create_one({"title": name})
+        return _names_repo.create_one(NameCreationRequestDTO(title=name))
 
     def _generate_random_word(self):
         """Generates a random word"""
@@ -37,3 +50,69 @@ class GenerateRandomName(Action):
             else:
                 word += random.choice(self._vowels)
         return word
+
+
+class CreateOneName(CreateOneAction):
+    """Create a single Name record in the repository"""
+
+    @property
+    def _repository(self) -> Repository:
+        return _names_repo
+
+
+class CreateManyNames(CreateManyAction):
+    """Create a list of name records in the repository"""
+
+    @property
+    def _repository(self) -> Repository:
+        return _names_repo
+
+
+class ReadOneName(ReadOneAction):
+    """Reads a single Name record from the repository"""
+
+    @property
+    def _repository(self) -> Repository:
+        return _names_repo
+
+
+class ReadManyNames(ReadManyAction):
+    """Reads a list of name records from the repository"""
+
+    @property
+    def _repository(self) -> Repository:
+        return _names_repo
+
+
+class UpdateOneName(UpdateOneAction):
+    """Updates a single Name record in the repository"""
+
+    @property
+    def _repository(self) -> Repository:
+        return _names_repo
+
+
+class UpdateManyNames(UpdateManyAction):
+    """Updates a list of name records in the repository"""
+
+    @property
+    def _repository(self) -> Repository:
+        return _names_repo
+
+
+class DeleteOneName(DeleteOneAction):
+    """Deletes a single Name record in the repository"""
+
+    @property
+    def _repository(self) -> Repository:
+        return _names_repo
+
+
+class DeleteManyNames(DeleteManyAction):
+    """
+    Deletes a list of name records in the repository basing on the Criteria passed
+    """
+
+    @property
+    def _repository(self) -> Repository:
+        return _names_repo